+++ packages/api-reference/src/components/Content/Schema/SchemaComposition.vue
@@ -74,7 +74,8 @@ const listboxOptions = computed((): ScalarListboxOption[] =>
   composition.value.map((schema, index: number) => {
     const resolved = resolve.schema(schema.original!)
     const label =
-      (getModelNameFromSchema(resolved) ?? getSchemaType(resolved)) || 'Schema'
+      (getModelNameFromSchema(resolved)?.label ?? getSchemaType(resolved)) ||
+      'Schema'
     return { id: String(index), label }
   }),
 )