+++ integrations/docker/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @scalarapi/docker-api-reference
 
+## 0.5.27
+
+## 0.5.26
+
 ## 0.5.25
 
 ## 0.5.24