+++ packages/api-client-react/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @scalar/api-client-react
 
+## 2.0.6
+
+## 2.0.5
+
 ## 2.0.4
 
 ### Patch Changes