+++ packages/api-reference/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @scalar/api-reference
 
+## 1.52.6
+
+## 1.52.5
+
 ## 1.52.4
 
 ### Patch Changes