+++ packages/server-side-rendering/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @scalar/server-side-rendering
 
+## 0.1.6
+
+## 0.1.5
+
 ## 0.1.4
 
 ### Patch Changes