+++ packages/types/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @scalar/types
 
+## 0.9.2
+
 ## 0.9.1
 
 ### Patch Changes