+++ packages/workspace-store/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @scalar/workspace-store
 
+## 0.46.3
+
+### Patch Changes
+
+- [#8952](https://github.com/scalar/scalar/pull/8952): Export shared Scalar custom header constants from `@scalar/helpers/http/scalar-headers` and consume them in request build/send flows.
+
 ## 0.46.2
 
 ### Patch Changes